requirements.txt
tensorflow>=2.20.0
keras>=3.11.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
pillow>=10.0.0
jupyter>=1.0.0
ipykernel>=6.0.0
scikeras>=0.13.0
pydot>=1.4.0
graphviz>=0.20.0
ipython>=8.0.0
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0
